cli-validation/python-cli/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "documentation-robotics"
version = "0.7.3"
description = "CLI tool for managing federated architecture data models"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "tinkermonkey"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "jsonschema>=4.17.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "python-dateutil>=2.8.0",
    "packaging>=23.0",
    # Template and graph dependencies
    "jinja2>=3.1.0",
    "networkx>=3.0",
    "jsonpath-ng>=1.5.0",
    "lxml>=4.9.0",
    "markdown>=3.4.0",
    # Interactive prompts
    "questionary>=2.0.0",
    # Text processing utilities
    "inflect>=7.0.0",
    # Web server for visualization
    "aiohttp>=3.9.0",
    "watchdog>=3.0.0",
    # AI chat integration
    "anthropic>=0.75.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "pre-commit>=3.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "types-PyYAML",
    "types-python-dateutil",
]

[project.scripts]
dr = "documentation_robotics.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
documentation_robotics = [
    "schemas/bundled/*.json",
    "schemas/bundled/.manifest.json",
    "schemas/bundled/common/*.json",
    "schemas/link-registry.json",
    "claude_integration/reference_sheets/*.md",
    "claude_integration/commands/*.md",
    "claude_integration/agents/*.md",
    "claude_integration/skills/*/SKILL.md",
    "claude_integration/templates/*.md",
    "claude_integration/templates/*.json",
    "claude_integration/templates/*.sh",
    "copilot_integration/agents/*.md",
    "copilot_integration/skills/*/SKILL.md",
    "viewer/dist/**/*",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=documentation_robotics --cov-report=term-missing"
asyncio_mode = "auto"
# Ignore unraisable exception warnings from asyncio event loop cleanup in watchdog library
# This is a known issue where subprocess transports are being garbage collected after their
# event loop has been closed, triggering BaseSubprocessTransport.__del__ exceptions.
# The fix (gc.collect() in FileMonitor.stop) helps but we need this filter as fallback.
filterwarnings = [
    "ignore::pytest.PytestUnraisableExceptionWarning:asyncio",
]

[tool.coverage.run]
source = ["documentation_robotics"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
# Ignore rules:
# E501: Line too long - Black handles line length, and E501 is not auto-fixable
# F821: Undefined name - Allow forward references in type hints (e.g., "ValidationResult", "networkx.DiGraph")
# N812: Lowercase imported as non-lowercase - Common convention for libraries like lxml (etree as ET)
ignore = ["E501", "F821", "N812"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
